+++ CHANGELOG.md
@@ -7,3 +7,18 @@ All notable changes to this project will be documented in this file.
 ## [0.0.1] - 2022-11-04
 
 Initial version of the extension, upload and inference of the resource
+
+
+## [0.0.2] - 2022-11-25
+
+Initial interface with metadata editing of the resource
+
+
+## [0.0.3] - 2022-11-09
+
+Updated the resource and package interface
+
+
+## [0.0.4] - 2022-12-13
+
+Updated installation instructions, improved the interface, and fixed bugs for resource and package update